Mostly fix subtle race condition around PRNG init
Google's PRNG fix can break due to a race condition when two instances
of PrngFixes or PRNGFixes (with different full class name) are running
at the same time.

They both will "fight" over who installs the
LinuxPRNGSecureRandomProvider provider class. Normally, each one will
succeed, check that they succeeded, and then continue on. However,
rarely, one will install its provider while the other is checking
that its installation succeeded, and cause the other to crash.

We now perform a looser check: that the provider's simple name is
merely "LinuxPRNGSecureRandomProvider" instead of checking
for an exact class.

This will fix any instances of the Tozny PrngFixes class from
throwing a SecurityException.

Sadly, though, we cannot figure out a foolproof way to prevent Google's
fix from crashing when the Tozny PrngFixes class initializes 100%
of the time. We have narrowed the race condition window to as
small as possible. But, if you know that other code is installing
a PRNG fix as well, then the best thing to do is to not use
any additional PRNG fix, Tozny's or otherwise.